fix: resolve u64::MAX overflow in SQL queries and test failures
- Fix u64::MAX overflow when cast to i64 for SQL bindings (caused empty
  query results when to_block was unspecified)
- Update test constants to match actual deployment block values
- Fix fractional rate test (capacity 2.0 -> 1.0 for correct behavior)
- Remove duplicate #[allow(dead_code)] attributes in rate_limiter.rs
- Rename config test to use AppConfig::default() instead of load()

+++ src/cache.rs
@@ -2004,7 +2004,8 @@ impl Cache {
         use crate::events::{PaymentChannelEventData, PaymentChannelEventType};
 
         let from = from_block.unwrap_or(0);
-        let to = to_block.unwrap_or(u64::MAX);
+        // Use i64::MAX to avoid overflow when casting to i64 for SQL binding
+        let to = to_block.unwrap_or(i64::MAX as u64);
 
         let events = match &self.pool {
             DatabasePool::Sqlite(pool) => {
